Add IOWebSocketChannel.innerWebSocket
Supersedes #141 - expose the entire `dart:io` web socket instead of
selectively forwarding more parts of the interface.

This does not cover the case of changing a setting immediately after the
socket is connected. As a followup we may want to add a
`Future<WebSocketChannel> get onConnected` to have a signal on
connection and enable immediately setting options.